ats-backend/requirements.txt
# Web framework
streamlit
google-generativeai
google-genai==1.31.0
flask==3.0.3




# Resume parsing
python-docx==1.1.0
pdfminer.six==20221105
PyPDF2==3.0.1
pdf2image==1.16.3

# NLP and matching
nltk==3.8.1
spacy==3.7.4
scikit-learn==1.4.2
sentence-transformers==2.6.1

# File handling
python-multipart==0.0.9

# Database
SQLAlchemy==2.0.42
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9  # for PostgreSQL (optional)
pinecone-client==6.0.0
pinecone==7.3.0
pgvector==0.4.1
protobuf==4.25.8

# CORS
flask-cors==4.0.1

# Environment variables
python-dotenv==1.0.1


# Optional: background tasks
celery==5.3.6
redis==5.0.4
